Restructure Flax examples module
